security - Moves entity enumerations into enums package.
