+++ src/main/java/com/entity/enums/ContractType.java
@@ -1,4 +1,4 @@
-package com.entity;
+package com.entity.enums;
 
 public enum ContractType {
