+++ src/main/java/com/entity/enums/MedicalRecord.java
@@ -1,4 +1,4 @@
-package com.entity;
+package com.entity.enums;
 
 public enum MedicalRecord {
