+++ src/main/java/com/entity/enums/MobileDeviceModel.java
@@ -1,4 +1,4 @@
-package com.entity;
+package com.entity.enums;
 
 public enum MobileDeviceModel {
     SAMSUNG,
